Fixed commands being accepted from other servers the bot is a member of

+++ discord_gambler/listeners/message_listeners.py
@@ -1,5 +1,6 @@
 from discord.ext import commands
 from decouple import config
+import logging
 from discord_gambler import _coinflip_channel, _guild_id
 import os
 
@@ -10,7 +11,13 @@ class MessageListeners(commands.Cog):
 
     @commands.Cog.listener()
     async def on_message(self, message):
-        if message.channel.name == _coinflip_channel:
+        logging.info(
+            f"Guild/Channel {message.channel.guild.id}/{message.channel.id}: {message.content}"
+        )
+        if (
+            message.channel.name == _coinflip_channel
+            and message.channel.guild.id == _guild_id
+        ):
             if message.author != self._bot.user:
                 if not message.content.startswith("!"):
                     try:
@@ -28,16 +35,17 @@ class MessageListeners(commands.Cog):
     # This event is called only if the command succeeded, i.e. all checks have passed and the user input it correctly.
     @commands.Cog.listener()
     async def on_command_completion(self, ctx):
-        try:
-            await ctx.message.delete()
-        except:
-            pass
+        if ctx.channel.name == _coinflip_channel and ctx.channel.guild.id == _guild_id:
+            try:
+                await ctx.message.delete()
+            except:
+                pass
 
     # An error handler that is called when an error is raised inside a command either through user input error, check failure, or an error in your own code.
     @commands.Cog.listener()
     async def on_command_error(self, ctx, error):
-        print(error)
-        try:
-            await ctx.message.delete()
-        except:
-            pass
+        if ctx.channel.name == _coinflip_channel and ctx.channel.guild.id == _guild_id:
+            try:
+                await ctx.message.delete()
+            except:
+                pass
